requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
Werkzeug==2.2.3
